Fix productImpression being called more than once
ShelfItem is being rerendered this cause the productImpression event to be triggered more than once when it shouldn't.

I didn't manage to avoid ShelfItem to be rerendered, so I changed to a place where the component changes less.
